Struct forest_json_utils::GoVecVisitor

pub struct GoVecVisitor<T, D = T> { /* fields omitted */ }

Helper visitor to match Go’s default behaviour of serializing uninitialized slices as null. This will be able to deserialize null as empty Vectors of the type.

T indicates the return type, and D is an optional generic to override the

Implementations

impl<T, D> GoVecVisitor<T, D>

pub fn new() -> Self